修正存檔 面向判定改採 矩陣處理 隨機函數 持需清查/異動 EX:      0 ~ 49 取 1個           1 ~ 100 取 2個 _random.nextInt(50)<1 == RandomArray100List()<3      1/50 發生機會           2/100 發生機會
git-svn-id: http://www.castiello-tsai.org:8523/svn/L1JTW@156 96aee97a-5c11-ed42-8518-d7f33e96f5f9

+++ src/l1j/server/server/types/UChar8.java
@@ -30,7 +30,7 @@ public class UChar8 {
 	public static char[] fromArray(long[] buff) {
 		char[] charBuff = new char[buff.length * 4];
 
-		for (int i = 0; i < buff.length; ++i) {
+		for (short i = 0; i < buff.length; ++i) {
 			charBuff[(i * 4) + 0] = (char) (buff[i] & 0xFF);
 			charBuff[(i * 4) + 1] = (char) ((buff[i] >> 8) & 0xFF);
 			charBuff[(i * 4) + 2] = (char) ((buff[i] >> 16) & 0xFF);
@@ -50,7 +50,7 @@ public class UChar8 {
 	public static char[] fromArray(byte[] buff) {
 		char[] charBuff = new char[buff.length];
 
-		for (int i = 0; i < buff.length; ++i) {
+		for (short i = 0; i < buff.length; ++i) {
 			charBuff[i] = (char) (buff[i] & 0xFF);
 		}
 
@@ -67,7 +67,7 @@ public class UChar8 {
 	public static char[] fromArray(byte[] buff, int length) {
 		char[] charBuff = new char[length];
 
-		for (int i = 0; i < length; ++i) {
+		for (short i = 0; i < length; ++i) {
 			charBuff[i] = (char) (buff[i] & 0xFF);
 		}
